producers/requirements.txt
websocket-client==1.6.4
kafka-python==2.0.2
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
